% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createWorldBankDataset}
\alias{createWorldBankDataset}
\title{Reshape the result from WorldBankAPI:::getWorldBankDataset function}
\usage{
createWorldBankDataset(df, colVar = "indicator.id",
  labelVar = "indicator.value", valueVar = "value")
}
\arguments{
\item{df}{}

\item{colVar}{}

\item{labelVar}{}

\item{valueVar}{}
}
\value{
data.table with the reshaped WB dataset
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

