% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWorldBankCountries}
\alias{getWorldBankCountries}
\title{Get a List of Countries, ISO code lookup, and region definitions}
\usage{
getWorldBankCountries(url = "http://api.worldbank.org", type = "countries")
}
\arguments{
\item{url}{URL of World Bank API}

\item{type}{}
}
\value{
data.table with country information
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

