% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWorldBankDataset}
\alias{getWorldBankDataset}
\title{Request a World Bank Dataset}
\usage{
getWorldBankDataset(url = "http://api.worldbank.org", source = "20",
  topic = NULL, countries = "all", MRV = "100", format = "json",
  per_page = "10000")
}
\arguments{
\item{url}{}

\item{source}{}

\item{topic}{}

\item{countries}{}

\item{MRV}{}

\item{format}{}

\item{per_page}{}

\item{frequency}{}
}
\value{
data.table with query result.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

