% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWorldBankIndicatorList}
\alias{getWorldBankIndicatorList}
\title{Get a list of available variables for a given WB data source}
\usage{
getWorldBankIndicatorList(url = "http://api.worldbank.org", source = "20",
  topic = NULL)
}
\arguments{
\item{url}{URL of World Bank API}

\item{source}{ID of the data source}

\item{topic}{Topic ID}
}
\value{
data.table with indicator descriptions
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

