% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWorldBankSeriesInfo}
\alias{getWorldBankSeriesInfo}
\title{Obtain additional infomration on individual data series}
\usage{
getWorldBankSeriesInfo(url = "http://api.worldbank.org", source = "20")
}
\arguments{
\item{url}{}

\item{source}{}
}
\value{
data.table with query results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

