% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWorldBankSources}
\alias{getWorldBankSources}
\title{Obtain a List of World Bank Data Sources}
\usage{
getWorldBankSources(url = "http://api.worldbank.org")
}
\arguments{
\item{url}{URL of World Bank API}
}
\value{
data.table with WB data sources and their descriptions
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

